Change usercontroller

+++ devitemapi/Controllers/Rbac/UserController.cs
@@ -1,6 +1,5 @@
 ﻿using AutoMapper;
 using devitemapi.Common;
-using devitemapi.Dto;
 using devitemapi.Entity;
 using devitemapi.Infrastructure.Repository.Interface;
 using devitemapi.Services.Interface;
@@ -12,7 +11,10 @@ using System;
 using System.Collections.Generic;
 using System.Data.SqlClient;
 using System.Linq;
+using System.Net;
 using System.Threading.Tasks;
+using devitemapi.Dto.User;
+using devitemapi.Infrastructure.ActionResult;
 
 namespace devitemapi.Controllers.Rbac
 {
@@ -27,7 +29,7 @@ namespace devitemapi.Controllers.Rbac
         private readonly IDevUserRepository _userRepository;
         private readonly IMapper _mapper;
 
-        public UserController(IUserService userService,IDevUserRepository repository,IMapper mapper)
+        public UserController(IUserService userService, IDevUserRepository repository, IMapper mapper)
         {
             this._userRepository = repository;
             this._mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
@@ -38,20 +40,18 @@ namespace devitemapi.Controllers.Rbac
         /// </summary>
         /// <param name="userId"></param>
         /// <returns></returns>
-        [HttpGet("{userId}")]
-        public async Task<IActionResult> GetUser(Guid userId)
+        [HttpGet("{userId:Guid}", Name = nameof(GetUser))]
+        [ProducesResponseType(typeof(UserDto), (int)HttpStatusCode.OK)]
+        [ProducesResponseType((int)HttpStatusCode.NotFound)]
+        public async Task<ActionResult<UserDto>> GetUser(Guid userId)
         {
-            ResponseDto response = new ResponseDto();
             var user = await _userRepository.GetUserAsync(userId);
-            var userDto = _mapper.Map<UserDto>(user);
-
-            if (userDto == null)
+            if (user == null)
             {
                 return NotFound();
             }
-
-            response.SetData(user);
-            return Ok(response);
+            var userDto = _mapper.Map<UserDto>(user);
+            return Ok(userDto);
         }
 
         /// <summary>
@@ -59,14 +59,12 @@ namespace devitemapi.Controllers.Rbac
         /// </summary>
         /// <returns></returns>
         [HttpGet]
-        //[ProducesResponseType(,)]   //声明可能返回的类型
-        public async Task<IActionResult> GetUsers(int limit = 10,int offest =0)
+        [ProducesResponseType(typeof(IEnumerable<UserDto>), (int)HttpStatusCode.OK)]   //声明可能返回的类型
+        public async Task<ActionResult<IEnumerable<UserDto>>> GetUsers(int limit = 10, int offest = 0)
         {
-            throw new ArgumentNullException(nameof(limit));
-            ResponseDto response = new ResponseDto();
+            return BadRequest(new ValidateErrorResult());
             var users = await _userRepository.GetUsersAsync();
             var userDto = _mapper.Map<IEnumerable<UserDto>>(users);
-            response.SetData(userDto);
             return Ok(userDto);
         }
 
@@ -76,13 +74,18 @@ namespace devitemapi.Controllers.Rbac
         /// <param name="user"></param>
         /// <returns></returns>
         [HttpPost]
-        public async Task<ResponseDto> CreateUser(UserDto user)
+        [ProducesResponseType((int)HttpStatusCode.Created)]
+        [ProducesResponseType((int)HttpStatusCode.BadRequest)]
+        public async Task<ActionResult> CreateUser([FromBody] UserAddDto user)
         {
-            ResponseDto response = new ResponseDto();
-            _userRepository.AddUser(_mapper.Map<DevUser>(user));
+            if (user == null)
+            {
+                return BadRequest();
+            }
+            var userDto = _mapper.Map<DevUser>(user);
+            _userRepository.AddUser(userDto);
             await _userRepository.SaveAsync();
-            return response;
-            
+            return CreatedAtRoute(nameof(GetUser), new { userId = userDto.Id }, null);
         }
 
         /// <summary>
@@ -91,20 +94,25 @@ namespace devitemapi.Controllers.Rbac
         /// <param name="userId">用户GUID</param>
         /// <returns></returns>
         [HttpDelete("{userId}")]
-        public async Task<IActionResult> DeleteUser(Guid userId)
+        [ProducesResponseType((int)HttpStatusCode.BadRequest)]
+        [ProducesResponseType((int)HttpStatusCode.NotFound)]
+        [ProducesResponseType((int)HttpStatusCode.NoContent)]
+        public async Task<ActionResult> DeleteUser(Guid userId)
         {
-
-            return NoContent();
-            ResponseDto response = new ResponseDto();
-            var user = await _userRepository.GetUserAsync(userId);
-            if(user == null)
+            if (userId == null || userId == Guid.Empty)
             {
-               response.SetFail("用户不存在");
-               return Ok(response);
+                return BadRequest();
             }
+
+            var user = await _userRepository.GetUserAsync(userId);
+            if (user == null)
+            {
+                return NotFound();
+            }
+
             _userRepository.DeleteUser(user);
             await _userRepository.SaveAsync();
-            return Ok(response);
+            return NoContent();
         }
 
         // /// <summary>
@@ -121,15 +129,30 @@ namespace devitemapi.Controllers.Rbac
         /// <summary>
         /// 修改用户
         /// </summary>
+        /// <param name="userId"></param>
         /// <param name="user"></param>
         /// <returns></returns>
         [HttpPut("userId")]
-        public async Task<ResponseDto> UpdateUser(Guid userId, DevUser user)
+        [ProducesResponseType((int)HttpStatusCode.BadRequest)]
+        [ProducesResponseType((int)HttpStatusCode.NotFound)]
+        [ProducesResponseType((int)HttpStatusCode.Created)]
+        public async Task<ActionResult> UpdateUser([FromRoute] Guid userId, [FromBody] UserUpdateDto user)
         {
-            ResponseDto response = new ResponseDto();
-            _userRepository.UpdateUser(user);
+            if (userId == null || userId == Guid.Empty || user is null)
+            {
+                return BadRequest();
+            }           
+            var userEntity = await _userRepository.GetUserAsync(userId);
+            if (userEntity == null)
+            {
+                return NotFound();
+            }
+
+            _mapper.Map(user,userEntity);            
+            _userRepository.UpdateUser(userEntity);
             await _userRepository.SaveAsync();
-            return response;
+
+            return CreatedAtRoute(nameof(GetUser), new { userId },null);
         }
     }
 }
